requirements.txt
PyDelphin
torch
transformers
datasets
wandb
smatch